# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/search": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Search for properties",
                "description": "Search for properties based on a keyword. The search is performed on the 'title' field of properties within the user's university (city).",
                "operationId": "2b1e0f7fab3082a6aefbc7f624b59ffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "keyword"
                                ],
                                "properties": {
                                    "keyword": {
                                        "type": "string",
                                        "example": "house"
                                    },
                                    "min_price": {
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "max_price": {
                                        "type": "integer",
                                        "example": 500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Search results"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Sunset Villa"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Rose Garden, NY"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 250000
                                                    },
                                                    "like": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "Fridge": {
                                                        "type": "string",
                                                        "example": "Yes"
                                                    },
                                                    "WaterTank": {
                                                        "type": "string",
                                                        "example": "Yes"
                                                    },
                                                    "Solar": {
                                                        "type": "string",
                                                        "example": "Available"
                                                    },
                                                    "Roommates": {
                                                        "type": "string",
                                                        "example": "2"
                                                    },
                                                    "wifi": {
                                                        "type": "string",
                                                        "example": "High Speed"
                                                    },
                                                    "room_image": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "https://example.com/storage/room.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No keyword provided"
                    },
                    "404": {
                        "description": "No properties found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/properties/bylocation": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get properties for authenticated user's university",
                "description": "Returns a list of properties filtered by the user's university (city).",
                "operationId": "getPropertiesByUniversity",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "location": {
                                                "type": "string",
                                                "example": "Zills Home"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Bearer token to access protected routes",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}